Specify database file at construction time.
The DB is an output of the task, so it should be associated with the
task object like the Butler is (also, allowing each run to put results
in a different association database rather defeats the purpose). This
design is also forward-compatible with the use of the Butler to manage
association output; just delete the extra argument.

Some hacking was needed to get around the fact that currently the DB
location must be specified by config, but that code can be removed
with no API changes once AssociationTask doesn't need config-time
specification.

As an immediate benefit, AssociationTask can now be treated as an
ordinary subtask.