Remove workaround for passing fullMetadata to ap_verify.
ap_verify has access to the task object at the moment, so it can query
the full metadata directly. Changes to its task management may require
support on the ap_pipe side in the future, but that's best deferred
until we have a clear picture of what ap_verify needs.