Implement PipelineTimingMetricTask.
This task allows the timing of a pipeline with well-defined "first"
and "last" tasks, i.e. one that's linear or close to linear. Unlike
lsst.verify.TimingMetricTask, the new task is not designed for general
use, and has some hardcoded assumptions about the pipeline (in
particular, that the first task is exposure-level and the last is
visit-level).