Deprecate processCcd subset.
The name "processCcd" is a throwback to the Gen 2 ProcessCcdTask, and
is not intuitive for new users who've only known Gen 3. The replacement
name apPipeSingleFrame conveys the term "single-frame processing"
(which is still in use) while distinguishing between the subset of
ApPipe and the standalone SingleFrame pipeline (which are conceptually
different enough that they must not be mixed up).